+++ src/views/Project.vue
@@ -419,8 +419,21 @@ export default {
       this.comments = await FirebaseService.SELECT_Comments(this.project_id)
     },
     DELETE_comment(comments, comment_index) {
-      // console.log(this.project_id)
-      FirebaseService.DELETE_comment(this.project_id, comments, comment_index)
+      this.$swal({
+       title: '정말 삭제하시겠습니까?',
+       text: "삭제된 댓글은 복구가 불가능합니다.",
+       type: 'warning',
+       showCancelButton: true,
+       confirmButtonColor: '#3085d6',
+       cancelButtonColor: '#d33',
+       confirmButtonText: '삭제',
+       cancelButtonText: '취소',
+      }).then((result) => {
+       if (result.value) {
+         this.$swal('Deleted!','댓글 삭제가 완료되었습니다.','success')
+         FirebaseService.DELETE_comment(this.project_id, comments, comment_index)
+       }
+     })
     },
     UPDATE_comment(comments, index) {
       var before = document.querySelector(`.before_${index}`)
